Chart-Prophet/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.56.0",
    "numpy>=2.4.0",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "sift-stack-py>=0.9.6",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "werkzeug>=3.1.4",
]
